requirements.txt
asgiref
Django
django-environ
sqlparse==0.4.2
djangorestframework
djangorestframework-simplejwt
python-dotenv
requests
Pillow
django-rest-passwordreset
sendgrid
django-allauth
cryptography
pyotp
psycopg2
gunicorn # for deployment purposes
whitenoise
django-password-validators
dj-database-url
django-cors-headers